requirements.txt
streamlit
tensorflow
pillow
pandas
openpyxl